// src/app/core/models/recursosHumanos/profile/recursosHumanos.dto.ts
import {UserDto} from '../../shared/user.dto';
import {DatosLaboralesDto} from '../../shared/datosLaborales.dto';

export interface RecursosHumanosDto {
  id: number;
  nombre: string;
  telefono: string;
  telefonoSecundario: string;
  imagenPerfil: string;
  user: UserDto;
  datosLaborales: DatosLaboralesDto;
}